moved methods to separate file, fixed imports
